feat!: SerialCircuit::decode_inplace and explicit option structs
