+++ tket-py/src/circuit/tk2circuit.rs
@@ -27,6 +27,7 @@ use serde::Serialize;
 use tket::circuit::CircuitHash;
 use tket::passes::pytket::lower_to_pytket;
 use tket::passes::CircuitChunks;
+use tket::serialize::pytket::{DecodeOptions, EncodeOptions};
 use tket::serialize::TKETDecode;
 use tket::{Circuit, TketOp};
 use tket_json_rs::circuit_json::SerialCircuit;
@@ -81,9 +82,12 @@ impl Tk2Circuit {
     /// Convert the [`Tk2Circuit`] to a tket1 circuit.
     pub fn to_tket1<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
         let circ = lower_to_pytket(&self.circ).convert_pyerrs()?;
-        SerialCircuit::encode_with_config(&circ, tket_qsystem::pytket::qsystem_encoder_config())
-            .convert_pyerrs()?
-            .to_tket1(py)
+        SerialCircuit::encode(
+            &circ,
+            EncodeOptions::new().with_config(tket_qsystem::pytket::qsystem_encoder_config()),
+        )
+        .convert_pyerrs()?
+        .to_tket1(py)
     }
 
     /// Apply a rewrite on the circuit.
@@ -154,9 +158,9 @@ impl Tk2Circuit {
         // Try to simplify tuple pack-unpack pairs, and other operations not supported by pytket.
         let circ = lower_to_pytket(&self.circ).convert_pyerrs()?;
         serde_json::to_string(
-            &SerialCircuit::encode_with_config(
+            &SerialCircuit::encode(
                 &circ,
-                tket_qsystem::pytket::qsystem_encoder_config(),
+                EncodeOptions::new().with_config(tket_qsystem::pytket::qsystem_encoder_config()),
             )
             .convert_pyerrs()?,
         )
@@ -170,7 +174,7 @@ impl Tk2Circuit {
     pub fn from_tket1_json(json: &str) -> PyResult<Self> {
         let circ = tket::serialize::load_tk1_json_str(
             json,
-            Some(tket_qsystem::pytket::qsystem_decoder_config()),
+            DecodeOptions::new().with_config(tket_qsystem::pytket::qsystem_decoder_config()),
         )
         .map_err(|e| {
             PyErr::new::<PyAttributeError, _>(format!("Could not load pytket circuit: {e}"))
@@ -183,9 +187,9 @@ impl Tk2Circuit {
         // Try to simplify tuple pack-unpack pairs, and other operations not supported by pytket.
         let circ = lower_to_pytket(&self.circ).convert_pyerrs()?;
         serde_json::to_vec(
-            &SerialCircuit::encode_with_config(
+            &SerialCircuit::encode(
                 &circ,
-                tket_qsystem::pytket::qsystem_encoder_config(),
+                EncodeOptions::new().with_config(tket_qsystem::pytket::qsystem_encoder_config()),
             )
             .convert_pyerrs()?,
         )
@@ -199,7 +203,7 @@ impl Tk2Circuit {
     pub fn from_tket1_json_bytes(json: &[u8]) -> PyResult<Self> {
         let circ = tket::serialize::load_tk1_json_reader(
             json,
-            Some(tket_qsystem::pytket::qsystem_decoder_config()),
+            DecodeOptions::new().with_config(tket_qsystem::pytket::qsystem_decoder_config()),
         )
         .map_err(|e| {
             PyErr::new::<PyAttributeError, _>(format!("Could not load pytket circuit: {e}"))
