+++ tket/src/serialize/pytket.rs
@@ -3,21 +3,22 @@
 mod config;
 pub mod decoder;
 pub mod encoder;
+mod error;
 pub mod extension;
+mod options;
 
 pub use config::{
     default_decoder_config, default_encoder_config, PytketDecoderConfig, PytketEncoderConfig,
     TypeTranslatorSet,
 };
 pub use encoder::PytketEncoderContext;
+pub use error::{OpConvertError, PytketDecodeError, PytketDecodeErrorInner, PytketEncodeError};
 pub use extension::PytketEmitter;
-
-use hugr::core::HugrNode;
+pub use options::{DecodeInsertionTarget, DecodeOptions, EncodeOptions};
 
 use hugr::hugr::hugrmut::HugrMut;
 use hugr::ops::handle::NodeHandle;
-use hugr::{Hugr, Wire};
-use itertools::Itertools;
+use hugr::{Hugr, Node};
 
 #[cfg(test)]
 mod tests;
@@ -25,17 +26,12 @@ mod tests;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::path::Path;
-use std::sync::Arc;
 use std::{fs, io};
 
-use hugr::ops::OpType;
-
-use derive_more::{Display, Error, From};
 use tket_json_rs::circuit_json::SerialCircuit;
 use tket_json_rs::register::{Bit, ElementId, Qubit};
 
 use crate::circuit::Circuit;
-use crate::serialize::pytket::extension::RegisterCount;
 
 use self::decoder::PytketDecoderContext;
 
@@ -64,113 +60,134 @@ pub trait TKETDecode: Sized {
     type EncodeError;
     /// Convert the serialized circuit to a circuit.
     ///
-    /// Uses a default set of extension decoders to translate operations.
-    fn decode(&self) -> Result<Circuit, Self::DecodeError>;
-    /// Convert the serialized circuit to a circuit.
-    fn decode_with_config(
+    /// See [DecodeOptions] to define the options used by the decoder.
+    ///
+    /// # Arguments
+    ///
+    /// - `options`: The options for the decoder.
+    ///
+    /// # Returns
+    ///
+    /// The encoded circuit.
+    fn decode(&self, options: DecodeOptions) -> Result<Circuit, Self::DecodeError>;
+    /// Convert the serialized circuit into a function definition in an existing HUGR.
+    ///
+    /// Does **not** modify the HUGR's entrypoint.
+    ///
+    /// # Arguments
+    ///
+    /// - `hugr`: The HUGR to define the function in.
+    /// - `target`: Where to insert the decoded circuit.
+    /// - `options`: The options for the decoder.
+    ///
+    /// # Returns
+    ///
+    /// The node id of the defined function.
+    fn decode_inplace(
         &self,
-        config: impl Into<Arc<PytketDecoderConfig>>,
-    ) -> Result<Circuit, Self::DecodeError>;
+        // This cannot be a generic HugrMut since it is stored inside the `PytketDecoderContext` that we to be Send+Sync
+        // (so that the extension decoder traits are dyn-compatible).
+        hugr: &mut Hugr,
+        target: DecodeInsertionTarget,
+        options: DecodeOptions,
+    ) -> Result<Node, Self::DecodeError>;
     /// Convert a circuit to a serialized pytket circuit.
     ///
-    /// Uses a default set of emitters to translate operations.
-    /// If the circuit contains non-std operations or types,
-    /// use [`TKETDecode::encode_with_config`] instead.
-    fn encode(circuit: &Circuit) -> Result<Self, Self::EncodeError>;
-    /// Convert a circuit to a serialized pytket circuit.
+    /// See [EncodeOptions] for the options used by the encoder.
     ///
-    /// You may use [`TKETDecode::encode`] if the circuit does not contain
-    /// non-std operations or types.
-    fn encode_with_config(
-        circuit: &Circuit,
-        config: impl Into<Arc<PytketEncoderConfig<Hugr>>>,
-    ) -> Result<Self, Self::EncodeError>;
+    /// # Arguments
+    ///
+    /// - `circuit`: The circuit to encode.
+    /// - `options`: The options for the encoder.
+    ///
+    /// # Returns
+    ///
+    /// A serialized pytket circuit.
+    fn encode(circuit: &Circuit, options: EncodeOptions) -> Result<Self, Self::EncodeError>;
 }
 
 impl TKETDecode for SerialCircuit {
     type DecodeError = PytketDecodeError;
     type EncodeError = PytketEncodeError;
 
-    fn decode(&self) -> Result<Circuit, Self::DecodeError> {
-        let config = default_decoder_config();
-        Self::decode_with_config(self, config)
-    }
-
-    fn decode_with_config(
-        &self,
-        config: impl Into<Arc<PytketDecoderConfig>>,
-    ) -> Result<Circuit, Self::DecodeError> {
+    fn decode(&self, options: DecodeOptions) -> Result<Circuit, Self::DecodeError> {
         let mut hugr = Hugr::new();
-
-        let mut decoder =
-            PytketDecoderContext::new(self, &mut hugr, None, None, Vec::new(), config)?;
-        decoder.run_decoder(&self.commands)?;
-        let main_func = decoder.finish()?;
-        hugr.set_entrypoint(main_func.node());
+        let main_func = self.decode_inplace(&mut hugr, DecodeInsertionTarget::Function, options)?;
+        hugr.set_entrypoint(main_func);
         Ok(hugr.into())
     }
 
-    fn encode(circuit: &Circuit) -> Result<Self, Self::EncodeError> {
-        let config = default_encoder_config();
-        Self::encode_with_config(circuit, config)
+    fn decode_inplace(
+        &self,
+        hugr: &mut Hugr,
+        _target: DecodeInsertionTarget,
+        options: DecodeOptions,
+    ) -> Result<Node, Self::DecodeError> {
+        let config = options
+            .config
+            .unwrap_or_else(|| default_decoder_config().into());
+
+        let mut decoder = PytketDecoderContext::new(
+            self,
+            hugr,
+            options.fn_name,
+            options.signature,
+            options.input_params,
+            config,
+        )?;
+        decoder.run_decoder(&self.commands)?;
+        Ok(decoder.finish()?.node())
     }
 
-    fn encode_with_config(
-        circuit: &Circuit,
-        config: impl Into<Arc<PytketEncoderConfig<Hugr>>>,
-    ) -> Result<Self, Self::EncodeError> {
-        let mut encoder = PytketEncoderContext::new(circuit, circuit.parent(), config)?;
-        encoder.run_encoder(circuit, circuit.parent())?;
-        let (serial, _) = encoder.finish(circuit, circuit.parent())?;
+    fn encode(circuit: &Circuit, options: EncodeOptions) -> Result<Self, Self::EncodeError> {
+        let config = options
+            .config
+            .unwrap_or_else(|| default_encoder_config().into());
+        let region = circuit.parent();
+        let mut encoder = PytketEncoderContext::new(circuit, region, config)?;
+        encoder.run_encoder(circuit, region)?;
+        let (serial, _) = encoder.finish(circuit, region)?;
         Ok(serial)
     }
 }
 
 /// Load a TKET1 circuit from a JSON file.
 ///
-/// If a decoder config is provided, it will be used to decode the circuit.
-/// Otherwise, it defaults to [`default_decoder_config`].
+/// See [DecodeOptions] for the options used by the decoder.
 pub fn load_tk1_json_file(
     path: impl AsRef<Path>,
-    config: Option<PytketDecoderConfig>,
+    options: DecodeOptions,
 ) -> Result<Circuit, PytketDecodeError> {
     let file = fs::File::open(path).map_err(PytketDecodeError::custom)?;
     let reader = io::BufReader::new(file);
-    load_tk1_json_reader(reader, config)
+    load_tk1_json_reader(reader, options)
 }
 
 /// Load a TKET1 circuit from a JSON reader.
 ///
-/// If a decoder config is provided, it will be used to decode the circuit.
-/// Otherwise, it defaults to [`default_decoder_config`].
+/// See [DecodeOptions] for the options used by the decoder.
 pub fn load_tk1_json_reader(
     json: impl io::Read,
-    config: Option<PytketDecoderConfig>,
+    options: DecodeOptions,
 ) -> Result<Circuit, PytketDecodeError> {
-    let config = config.unwrap_or_else(default_decoder_config);
     let ser: SerialCircuit = serde_json::from_reader(json).map_err(PytketDecodeError::custom)?;
-    let circ: Circuit = ser.decode_with_config(config)?;
+    let circ: Circuit = ser.decode(options)?;
     Ok(circ)
 }
 
 /// Load a TKET1 circuit from a JSON string.
 ///
-/// If a decoder config is provided, it will be used to decode the circuit.
-/// Otherwise, it defaults to [`default_decoder_config`].
-pub fn load_tk1_json_str(
-    json: &str,
-    config: Option<PytketDecoderConfig>,
-) -> Result<Circuit, PytketDecodeError> {
+/// See [DecodeOptions] for the options used by the decoder.
+pub fn load_tk1_json_str(json: &str, options: DecodeOptions) -> Result<Circuit, PytketDecodeError> {
     let reader = json.as_bytes();
-    load_tk1_json_reader(reader, config)
+    load_tk1_json_reader(reader, options)
 }
 
 /// Save a circuit to file in TK1 JSON format.
 ///
 /// You may need to normalize the circuit using [`lower_to_pytket`] before saving.
 ///
-/// If an encoder config is provided, it will be used to encode the circuit.
-/// Otherwise, it defaults to [`default_encoder_config`].
+/// See [EncodeOptions] for the options used by the encoder.
 ///
 /// # Errors
 ///
@@ -179,19 +196,18 @@ pub fn load_tk1_json_str(
 pub fn save_tk1_json_file(
     circ: &Circuit,
     path: impl AsRef<Path>,
-    config: Option<PytketEncoderConfig<Hugr>>,
+    options: EncodeOptions,
 ) -> Result<(), PytketEncodeError> {
     let file = fs::File::create(path).map_err(PytketEncodeError::custom)?;
     let writer = io::BufWriter::new(file);
-    save_tk1_json_writer(circ, writer, config)
+    save_tk1_json_writer(circ, writer, options)
 }
 
 /// Save a circuit in TK1 JSON format to a writer.
 ///
 /// You may need to normalize the circuit using [`lower_to_pytket`] before saving.
 ///
-/// If an encoder config is provided, it will be used to encode the circuit.
-/// Otherwise, it defaults to [`default_encoder_config`].
+/// See [EncodeOptions] for the options used by the encoder.
 ///
 /// # Errors
 ///
@@ -200,10 +216,9 @@ pub fn save_tk1_json_file(
 pub fn save_tk1_json_writer(
     circ: &Circuit,
     w: impl io::Write,
-    config: Option<PytketEncoderConfig<Hugr>>,
+    options: EncodeOptions,
 ) -> Result<(), PytketEncodeError> {
-    let config = config.unwrap_or_else(default_encoder_config);
-    let serial_circ = SerialCircuit::encode_with_config(circ, config)?;
+    let serial_circ = SerialCircuit::encode(circ, options)?;
     serde_json::to_writer(w, &serial_circ).map_err(PytketEncodeError::custom)?;
     Ok(())
 }
@@ -212,8 +227,7 @@ pub fn save_tk1_json_writer(
 ///
 /// You may need to normalize the circuit using [`lower_to_pytket`] before saving.
 ///
-/// If an encoder config is provided, it will be used to encode the circuit.
-/// Otherwise, it defaults to [`default_encoder_config`].
+/// See [EncodeOptions] for the options used by the encoder.
 ///
 /// # Errors
 ///
@@ -221,375 +235,12 @@ pub fn save_tk1_json_writer(
 /// supported by pytket.
 pub fn save_tk1_json_str(
     circ: &Circuit,
-    config: Option<PytketEncoderConfig<Hugr>>,
+    options: EncodeOptions,
 ) -> Result<String, PytketEncodeError> {
     let mut buf = io::BufWriter::new(Vec::new());
-    save_tk1_json_writer(circ, &mut buf, config)?;
+    save_tk1_json_writer(circ, &mut buf, options)?;
     let bytes = buf.into_inner().unwrap();
     String::from_utf8(bytes).map_err(PytketEncodeError::custom)
-}
-
-/// Error type for conversion between pytket operations and tket ops.
-#[derive(Display, derive_more::Debug, Error)]
-#[non_exhaustive]
-#[debug(bounds(N: HugrNode))]
-pub enum OpConvertError<N = hugr::Node> {
-    /// Tried to decode a tket1 operation with not enough parameters.
-    #[display(
-        "Operation {} is missing encoded parameters. Expected at least {expected} but only \"{}\" were specified.",
-        optype,
-        params.iter().join(", "),
-    )]
-    MissingSerialisedParams {
-        /// The operation name.
-        optype: OpType,
-        /// The expected number of parameters.
-        expected: usize,
-        /// The given of parameters.
-        params: Vec<String>,
-    },
-    /// Tried to decode a tket1 operation with not enough qubit/bit arguments.
-    #[display(
-        "Operation {} is missing encoded arguments. Expected {expected_qubits} and {expected_bits}, but only \"{args:?}\" were specified.",
-        optype,
-    )]
-    MissingSerialisedArguments {
-        /// The operation name.
-        optype: OpType,
-        /// The expected number of qubits.
-        expected_qubits: usize,
-        /// The expected number of bits.
-        expected_bits: usize,
-        /// The given of parameters.
-        args: Vec<ElementId>,
-    },
-    /// Tried to query the values associated with an unexplored wire.
-    ///
-    /// This reflects a bug in the operation encoding logic of an operation.
-    #[display("Could not find values associated with wire {wire}.")]
-    WireHasNoValues {
-        /// The wire that has no values.
-        wire: Wire<N>,
-    },
-    /// Tried to add values to an already registered wire.
-    ///
-    /// This reflects a bug in the operation encoding logic of an operation.
-    #[display("Tried to register values for wire {wire}, but it already has associated values.")]
-    WireAlreadyHasValues {
-        /// The wire that already has values.
-        wire: Wire<N>,
-    },
-}
-
-/// Error type for conversion between tket ops and pytket operations.
-#[derive(derive_more::Debug, Display, Error, From)]
-#[non_exhaustive]
-#[debug(bounds(N: HugrNode))]
-pub enum PytketEncodeError<N = hugr::Node> {
-    /// Tried to encode a non-dataflow region.
-    #[display("Cannot encode non-dataflow region at {region} with type {optype}.")]
-    NonDataflowRegion {
-        /// The region that is not a dataflow region.
-        region: N,
-        /// The operation type of the region.
-        optype: String,
-    },
-    /// Operation conversion error.
-    #[from]
-    OpConversionError(OpConvertError<N>),
-    /// Custom user-defined error raised while encoding an operation.
-    #[display("Error while encoding operation: {msg}")]
-    CustomError {
-        /// The custom error message
-        msg: String,
-    },
-}
-
-impl<N> PytketEncodeError<N> {
-    /// Create a new error with a custom message.
-    pub fn custom(msg: impl ToString) -> Self {
-        Self::CustomError {
-            msg: msg.to_string(),
-        }
-    }
-}
-
-/// Error type for conversion between tket2 ops and pytket operations.
-#[derive(derive_more::Debug, Display, Error, Clone)]
-#[non_exhaustive]
-#[display(
-    "{inner}{context}",
-    context = {
-        match (pytket_op, hugr_op) {
-            (Some(pytket_op), Some(hugr_op)) => format!(". While decoding a pytket {pytket_op} as a hugr {hugr_op}"),
-            (Some(pytket_op), None) => format!(". While decoding a pytket {pytket_op}"),
-            (None, Some(hugr_op)) => format!(". While decoding a hugr {hugr_op}"),
-            (None, None) => String::new(),
-        }
-    },
-)]
-pub struct PytketDecodeError {
-    /// The kind of error.
-    pub inner: PytketDecodeErrorInner,
-    /// The pytket operation that caused the error, if applicable.
-    pub pytket_op: Option<String>,
-    /// The hugr operation that caused the error, if applicable.
-    pub hugr_op: Option<String>,
-}
-
-impl PytketDecodeError {
-    /// Create a new error with a custom message.
-    pub fn custom(msg: impl ToString) -> Self {
-        PytketDecodeErrorInner::CustomError {
-            msg: msg.to_string(),
-        }
-        .into()
-    }
-
-    /// Create an error for an unknown qubit register.
-    pub fn unknown_qubit_reg(register: &tket_json_rs::register::ElementId) -> Self {
-        PytketDecodeErrorInner::UnknownQubitRegister {
-            register: register.to_string(),
-        }
-        .into()
-    }
-
-    /// Create an error for an unknown bit register.
-    pub fn unknown_bit_reg(register: &tket_json_rs::register::ElementId) -> Self {
-        PytketDecodeErrorInner::UnknownBitRegister {
-            register: register.to_string(),
-        }
-        .into()
-    }
-
-    /// Add the pytket operation name to the error.
-    pub fn pytket_op(mut self, op: &tket_json_rs::OpType) -> Self {
-        self.pytket_op = Some(format!("{op}"));
-        self
-    }
-
-    /// Add the hugr operation name to the error.
-    pub fn hugr_op(mut self, op: impl ToString) -> Self {
-        self.hugr_op = Some(op.to_string());
-        self
-    }
-}
-
-impl From<PytketDecodeErrorInner> for PytketDecodeError {
-    fn from(inner: PytketDecodeErrorInner) -> Self {
-        Self {
-            inner,
-            pytket_op: None,
-            hugr_op: None,
-        }
-    }
-}
-
-/// Error variants of [`PytketDecodeError`], signalling errors during the
-/// conversion between tket2 ops and pytket operations.
-#[derive(derive_more::Debug, Display, Error, Clone)]
-#[non_exhaustive]
-pub enum PytketDecodeErrorInner {
-    /// The pytket circuit uses multi-indexed registers.
-    //
-    // This could be supported in the future, if there is a need for it.
-    #[display("Register {register} in the circuit has multiple indices. Tket2 does not support multi-indexed registers")]
-    MultiIndexedRegister {
-        /// The register name.
-        register: String,
-    },
-    /// Found an unexpected register name.
-    #[display("Found an unknown qubit register name: {register}")]
-    UnknownQubitRegister {
-        /// The unknown register name.
-        register: String,
-    },
-    /// Found an unexpected bit register name.
-    #[display("Found an unknown bit register name: {register}")]
-    UnknownBitRegister {
-        /// The unknown register name.
-        register: String,
-    },
-    /// The given signature to use for the HUGR's input wires is not compatible with the number of qubits and bits in the pytket circuit.
-    ///
-    /// The expected number of qubits and bits may be different depending on the [`PytketTypeTranslator`][extension::PytketTypeTranslator]s used in the decoder config.
-    #[display(
-        "The given input types {input_types} to use for the HUGR's input wires are not compatible with the number of qubits and bits in the pytket circuit. Expected {expected_qubits} qubits and {expected_bits} bits, but found {circ_qubits} qubits and {circ_bits} bits",
-        input_types = input_types.iter().join(", "),
-    )]
-    InvalidInputSignature {
-        /// The given input types.
-        input_types: Vec<String>,
-        /// The expected number of qubits in the signature.
-        expected_qubits: usize,
-        /// The expected number of bits in the signature.
-        expected_bits: usize,
-        /// The number of qubits in the pytket circuit.
-        circ_qubits: usize,
-        /// The number of bits in the pytket circuit.
-        circ_bits: usize,
-    },
-    /// The signature to use for the HUGR's output wires is not compatible with the number of qubits and bits in the pytket circuit.
-    ///
-    /// We don't do any kind of type conversion, so this depends solely on the last operation to update each register.
-    #[display(
-        "The expected output types {expected_types} are not compatible with the actual output types {actual_types}, obtained from decoding the pytket circuit",
-        expected_types = expected_types.iter().join(", "),
-        actual_types = actual_types.iter().join(", "),
-    )]
-    InvalidOutputSignature {
-        /// The expected types of the input wires.
-        expected_types: Vec<String>,
-        /// The actual types of the input wires.
-        actual_types: Vec<String>,
-    },
-    /// A pytket operation had some input registers that couldn't be mapped to hugr wires.
-    //
-    // Some of this errors will be avoided in the future once we are able to decompose complex types automatically.
-    #[display(
-        "Could not find a wire with the required qubit arguments [{qubit_args:?}] and bit arguments [{bit_args:?}]",
-        qubit_args = qubit_args.iter().join(", "),
-        bit_args = bit_args.iter().join(", "),
-    )]
-    ArgumentCouldNotBeMapped {
-        /// The qubit arguments that couldn't be mapped.
-        qubit_args: Vec<String>,
-        /// The bit arguments that couldn't be mapped.
-        bit_args: Vec<String>,
-    },
-    /// Found an unexpected number of input wires when decoding an operation.
-    #[display(
-            "Expected {expected_values} input value wires{expected_types} and {expected_params} input parameters, but found {actual_values} values{actual_types} and {actual_params} parameters",
-            expected_types = match expected_types {
-                None => "".to_string(),
-                Some(tys) => format!(" with types [{}]", tys.iter().join(", ")),
-            },
-            actual_types = match actual_types {
-                None => "".to_string(),
-                Some(tys) => format!(" with types [{}]", tys.iter().join(", ")),
-            },
-        )]
-    UnexpectedInputWires {
-        /// The expected amount of input wires.
-        expected_values: usize,
-        /// The expected amount of input parameters.
-        expected_params: usize,
-        /// The actual amount of input wires.
-        actual_values: usize,
-        /// The actual amount of input parameters.
-        actual_params: usize,
-        /// The expected types of the input wires.
-        expected_types: Option<Vec<String>>,
-        /// The actual types of the input wires.
-        actual_types: Option<Vec<String>>,
-    },
-    /// Found an unexpected input type when decoding an operation.
-    #[display(
-        "Found an unexpected type {unknown_type} in the input wires, in input signature ({all_types})",
-        all_types = all_types.iter().join(", "),
-    )]
-    UnexpectedInputType {
-        /// The unknown type.
-        unknown_type: String,
-        /// All the input types specified for the operation.
-        all_types: Vec<String>,
-    },
-    /// Tried to track the output wires of a node, but the number of tracked elements didn't match the ones in the output wires.
-    #[display(
-        "Tried to track the output wires of a node, but the number of tracked elements didn't match the ones in the output wires. Expected {expected_qubits} qubits and {expected_bits} bits, but found {circ_qubits} qubits and {circ_bits} bits in the node outputs"
-    )]
-    UnexpectedNodeOutput {
-        /// The expected number of qubits.
-        expected_qubits: usize,
-        /// The expected number of bits.
-        expected_bits: usize,
-        /// The number of qubits in HUGR node outputs.
-        circ_qubits: usize,
-        /// The number of bits in HUGR node output.
-        circ_bits: usize,
-    },
-    /// Custom user-defined error raised while encoding an operation.
-    #[display("Error while decoding operation: {msg}")]
-    CustomError {
-        /// The custom error message
-        msg: String,
-    },
-    /// Input parameter was defined multiple times.
-    #[display("Parameter {param} was defined multiple times in the input signature")]
-    DuplicatedParameter {
-        /// The parameter name.
-        param: String,
-    },
-    /// Not enough parameter names given for the input signature.
-    #[display("Tried to initialize a pytket circuit decoder with {num_params_given} given parameter names, but more were required by the input signature")]
-    MissingParametersInInput {
-        /// The number of parameters given.
-        num_params_given: usize,
-    },
-    /// We don't support complex types containing parameters in the input.
-    //
-    // This restriction may be relaxed in the future.
-    #[display("Complex type {ty} contains {num_params} inside it. We only support input parameters in standalone 'float' or 'rotation'-typed wires")]
-    UnsupportedParametersInInput {
-        /// The type that contains the parameters.
-        ty: String,
-        /// The number of parameters in the type.
-        num_params: usize,
-    },
-    /// We couldn't find a wire that contains the required type.
-    #[display(
-        "Could not find a wire with type {ty} that contains {expected_arguments}",
-        expected_arguments = match (qubit_args.is_empty(), bit_args.is_empty()) {
-            (true, true) => "no arguments".to_string(),
-            (true, false) => format!("pytket bit arguments [{}]", bit_args.iter().join(", ")),
-            (false, true) => format!("pytket qubit arguments [{}]", qubit_args.iter().join(", ")),
-            (false, false) => format!("pytket qubit and bit arguments [{}] and [{}]", qubit_args.iter().join(", "), bit_args.iter().join(", ")),
-        },
-    )]
-    NoMatchingWire {
-        /// The type that couldn't be found.
-        ty: String,
-        /// The qubit registers expected in the wire.
-        qubit_args: Vec<String>,
-        /// The bit registers expected in the wire.
-        bit_args: Vec<String>,
-    },
-    /// The number of pytket registers expected for an operation is not enough.
-    ///
-    /// This is usually caused by a mismatch between the input signature and the number of registers in the pytket circuit.
-    ///
-    /// The expected number of registers may be different depending on the [`PytketTypeTranslator`][extension::PytketTypeTranslator]s used in the decoder config.
-    #[display(
-        "Expected {expected_count} to map types ({expected_types}), but only got {actual_count}",
-        expected_types = expected_types.iter().join(", "),
-    )]
-    NotEnoughPytketRegisters {
-        /// The types we tried to get wires for.
-        expected_types: Vec<String>,
-        /// The number of registers required by the types.
-        expected_count: RegisterCount,
-        /// The number of registers we actually got.
-        actual_count: RegisterCount,
-    },
-    /// A qubit was marked as outdated, but was expected to be fresh.
-    #[display("Discarded qubit {qubit} cannot be used as an input")]
-    OutdatedQubit {
-        /// The qubit that was marked as outdated.
-        qubit: String,
-    },
-    /// A bit was marked as outdated, but was expected to be fresh.
-    #[display("Discarded bit {bit} cannot be used as an input")]
-    OutdatedBit {
-        /// The bit that was marked as outdated.
-        bit: String,
-    },
-}
-
-impl PytketDecodeErrorInner {
-    /// Wrap the error in a [`PytketDecodeError`].
-    pub fn wrap(self) -> PytketDecodeError {
-        PytketDecodeError::from(self)
-    }
 }
 
 /// A hashed register, used to identify registers in the [`Tk1Decoder::register_wire`] map,
