+++ tket/src/serialize/pytket/tests.rs
@@ -21,6 +21,7 @@ use crate::circuit::Circuit;
 use crate::extension::rotation::{rotation_type, ConstRotation, RotationOp};
 use crate::extension::sympy::SympyOpDef;
 use crate::extension::TKET1_EXTENSION_ID;
+use crate::serialize::pytket::{DecodeOptions, EncodeOptions};
 use crate::TketOp;
 
 const SIMPLE_JSON: &str = r#"{
@@ -442,14 +443,14 @@ fn json_roundtrip(
     let ser: circuit_json::SerialCircuit = serde_json::from_str(circ_s).unwrap();
     assert_eq!(ser.commands.len(), num_commands);
 
-    let circ: Circuit = ser.decode().unwrap();
+    let circ: Circuit = ser.decode(DecodeOptions::new()).unwrap();
     assert_eq!(circ.qubit_count(), num_qubits);
 
     if !has_tk1_ops {
         check_no_tk1_ops(&circ);
     }
 
-    let reser: SerialCircuit = SerialCircuit::encode(&circ).unwrap();
+    let reser: SerialCircuit = SerialCircuit::encode(&circ, EncodeOptions::new()).unwrap();
     validate_serial_circ(&reser);
     compare_serial_circs(&ser, &reser);
 }
@@ -460,11 +461,11 @@ fn json_roundtrip(
 fn json_file_roundtrip(#[case] circ: impl AsRef<std::path::Path>) {
     let reader = BufReader::new(std::fs::File::open(circ).unwrap());
     let ser: circuit_json::SerialCircuit = serde_json::from_reader(reader).unwrap();
-    let circ: Circuit = ser.decode().unwrap();
+    let circ: Circuit = ser.decode(DecodeOptions::new()).unwrap();
 
     check_no_tk1_ops(&circ);
 
-    let reser: SerialCircuit = SerialCircuit::encode(&circ).unwrap();
+    let reser: SerialCircuit = SerialCircuit::encode(&circ, EncodeOptions::new()).unwrap();
     validate_serial_circ(&reser);
     compare_serial_circs(&ser, &reser);
 }
@@ -477,8 +478,11 @@ fn json_file_roundtrip(#[case] circ: impl AsRef<std::path::Path>) {
 #[case::preset_qubits(circ_preset_qubits(), Signature::new_endo(vec![qb_t(), qb_t(), qb_t()]))]
 #[case::preset_parameterized(circ_parameterized(), Signature::new(vec![qb_t(), rotation_type(), rotation_type(), rotation_type()], vec![qb_t()]))]
 fn circuit_roundtrip(#[case] circ: Circuit, #[case] decoded_sig: Signature) {
-    let ser: SerialCircuit = SerialCircuit::encode(&circ).unwrap();
-    let deser: Circuit = ser.decode().unwrap();
+    let ser: SerialCircuit =
+        SerialCircuit::encode(&circ, EncodeOptions::new()).unwrap_or_else(|e| panic!("{e}"));
+    let deser: Circuit = ser
+        .decode(DecodeOptions::new())
+        .unwrap_or_else(|e| panic!("{e}"));
 
     let deser_sig = deser.circuit_signature();
     assert_eq!(
@@ -492,7 +496,7 @@ fn circuit_roundtrip(#[case] circ: Circuit, #[case] decoded_sig: Signature) {
         &decoded_sig, &deser_sig
     );
 
-    let reser = SerialCircuit::encode(&deser).unwrap();
+    let reser = SerialCircuit::encode(&deser, EncodeOptions::new()).unwrap();
     validate_serial_circ(&reser);
     compare_serial_circs(&ser, &reser);
 }
@@ -509,13 +513,13 @@ fn circuit_roundtrip(#[case] circ: Circuit, #[case] decoded_sig: Signature) {
 fn test_add_angle_serialise(#[case] circ_add_angles: (Circuit, String)) {
     let (circ, expected) = circ_add_angles;
 
-    let ser: SerialCircuit = SerialCircuit::encode(&circ).unwrap();
+    let ser: SerialCircuit = SerialCircuit::encode(&circ, EncodeOptions::new()).unwrap();
     assert_eq!(ser.commands.len(), 1);
     assert_eq!(ser.commands[0].op.op_type, optype::OpType::Rx);
     assert_eq!(ser.commands[0].op.params, Some(vec![expected]));
 
-    let deser: Circuit = ser.decode().unwrap();
-    let reser = SerialCircuit::encode(&deser).unwrap();
+    let deser: Circuit = ser.decode(DecodeOptions::new()).unwrap();
+    let reser = SerialCircuit::encode(&deser, EncodeOptions::new()).unwrap();
     validate_serial_circ(&reser);
     compare_serial_circs(&ser, &reser);
 }
